Версия 2021.02.19

+++ SCHEMA/public/FUNCTION/changed_inventory.sql
@@ -25,7 +25,7 @@ begin
 			rgoods.amount = rgoods.amount - ap.amount;
 		
 			insert into balance_goods (owner_id, document_date, document_name, document_number, reference_id, amount, operation_summa)
-				values (new.id, new.doc_date, kind_name, new.doc_number, rgoods.goods_id, rgoods.amount, abs(goods_price * rgoods.amount)) returning id into b_id;
+				values (new.id, new.doc_date, kind_name, new.doc_number, rgoods.goods_id, rgoods.amount, abs(goods_price::numeric * rgoods.amount)) returning id into b_id;
 			update balance_goods
 				set status_id = 1112
 				where id = b_id;
